test(Uploader): add demo test
