+++ packages/uploader/demo/index.json
@@ -1,0 +1,8 @@
+{
+  "component": true,
+  "usingComponents": {
+    "van-button": "../../button/index",
+    "van-uploader": "../../uploader/index",
+    "demo-block": "../../../example/components/demo-block/index"
+  }
+}
